Use SocketMode for connections
Updates Slacker to use the Events API over the SocketMode communication
protocol, replacing calls to the deprecated RTM API.

A few key functions have changed, notably the NewClient function now
requires both an App and Bot token to be specified. In addition, the
typing indicator is not supported in SocketMode and has been removed
altogether.