Remove app_mentions from manifest
The manifest we ship sets up both message.im and message.channel in
addition to app_mention, but if app_mention is included with either
message event types then in the case where an app_mention is in a dm
(message.im) or channel (message.channel) that the bot is in, the bot
will receive two events and respond twice (as was seen in #82).

While a user may want that for some reason, they would also need to
account for that in their command code. To be friendlier for users
attempting to get started, we'll remove the app_mention since message.im
and message.channel will send an event whether the bot is mentioned or
not as long as it is in the channel.